fix warning when building for win32

+++ tiny_gltf.h
@@ -3219,7 +3219,7 @@ bool ReadWholeFile(std::vector<unsigned char> *out, std::string *err,
     return false;
   }
 
-  out->resize(sz);
+  out->resize(static_cast<size_t>(sz));
   f.read(reinterpret_cast<char *>(&out->at(0)),
          static_cast<std::streamsize>(sz));
